move JSON into tutor/src/tours
