+++ tutor/src/models/tour/region.js
@@ -4,7 +4,7 @@ import {
 import { extend } from 'lodash';
 
 // contains the mappings from region-ids to audiance tags
-import RegionData from './regions.json';
+import RegionData from '../../tours/regions.json';
 
 // TourRegion
 // Wraps an area of the screen, maps it's id to a given set of audience tags
@@ -12,19 +12,18 @@ import RegionData from './regions.json';
 @identifiedBy('tour/region')
 export default class TourRegion extends BaseModel {
 
-  static forIdentifier(id) {
-    return RegionData[id] ? new TourRegion(
-      extend({ id: id }, RegionData[id])
-    ) : undefined;
+  static forIdentifier(id, options = {}) {
+    const data = RegionData[id] || RegionData[options.fallback];
+    return data ? new TourRegion(extend({ id: id }, data[id])) : undefined;
   }
 
   @identifier id;
   @field courseId;
 
+  @field({ type: 'array' }) tour_ids;
+
   @computed get domSelector() {
     return `[data-tour-region-id="${this.id}"]`;
   }
-
-  @field({ type: 'array' }) tour_ids;
 
 }
