--- tutor/src/models/tour/regions.json
@@ -1,5 +1,0 @@
-{
-  "teacher-calendar": {
-    "tour_ids": ["teach-new-preview"]
-  }
-}
