+++ tutor/src/tours/add-homework.json
@@ -1,0 +1,43 @@
+{
+  "id": "add-homework",
+  "name": "Add a Homework using the builder",
+  "audience_tags": [
+    "teacher"
+  ],
+  "steps": [
+    {
+      "title": "OpenStax Tutor homework",
+      "cotent": "A homework assignment includes questions you assign and up to four personalized questions assigned by OpenStax Tutor.",
+      "position": "center"
+    }, {
+        "title": "Build a homework",
+        "content": "Ready to make a homework assignment? Fill in this form and click “select problems.”"
+    }, {
+        "title": "First, select a chapter",
+        "body": "Select a chapter you want to assign, then scroll down and click “show problems.” Uncheck any sections you don’t want to cover."
+    }, {
+        "title": "Personalized questions",
+        "body": "OpenStax Tutor generates questions using machine learning algorithms that study a student’s performance. You can allow up to four, but no fewer than two personalized questions per homework assignment.",
+        "anchor": {
+            "id": "sections-button"
+        }
+    }, {
+        "body": "[best-practices-icon] Don’t forget—exclude questions you don’t want your students to get before you publish an assignment. Questions can’t be removed after it’s published.",
+        "anchor": {
+            "id": "sections-button"
+        }
+    }, {
+        "title": "How OpenStax Tutor selects questions",
+        "body": "Based on sections you assign and student performance, OpenStax Tutor selects questions that will help students improve their understanding of [the section or topic, depending on how we’re using those terms].",
+        "anchor": {
+            "id": "sections-button"
+        }
+    }, {
+        "title": "Add questions",
+        "body": "To add a homework question, hover over the question and click “Add question.”",
+        "anchor": {
+            "id": "exercise-preview"
+        }
+    }
+  ]
+}
