+++ tutor/src/tours/index.js
@@ -1,0 +1,8 @@
+/* global require:true */
+import { extend } from 'lodash';
+
+export default {
+
+  'teach-new-preview': require('./teacher-calendar.json'),
+  'add-homework': require('./add-homework.json'),
+};
